+++ .eslintrc.js
@@ -71,6 +71,8 @@ module.exports = {
             'jest/require-top-level-describe': 'off',
             'jest/valid-describe': 'off',
             'jest/no-hooks': 'off',
+            'jest/prefer-called-with': 'off',
+            'jest/prefer-spy-on': 'off',
         }
     }],
     settings: {